requirements.txt

python-telegram-bot==21.5
openai>=1.50.0
requests
Pillow
python-dotenv
